requirements.txt
pika==latest
python-dotenv==latest
pandas==latest
pydantic==latest
pytest==latest
python-dateutil==latest